Now Wired can be compiled on win32. Added win32 related stuff, fixed some compiler warnings, some code unsupported on win32 were put between pre-compiling directives and some files were renamed to prevent object files with the same name.
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk/wired@1408 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/engine/Settings.cpp
@@ -155,7 +155,7 @@ void Settings::Load()
 
   // audio part, options
   conf->SetPath(wxT("/AudioOptions"));
-  conf->Read(wxT("SampleRate"), &SampleRate, DEFAULT_SAMPLE_RATE);
+  conf->Read(wxT("SampleRate"), &SampleRate, (int)DEFAULT_SAMPLE_RATE);
   conf->Read(wxT("SampleFormat"), &SampleFormat, 0);
   conf->Read(wxT("SamplesPerBuffer"), &SamplesPerBuffer, DEFAULT_SAMPLES_PER_BUFFER);
